pyproject.toml
[project]
name = "equityai-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beanie>=2.0.0",
    "beautifulsoup4>=4.13.4",
    "easy-equities-client>=0.5.0",
    "fastapi>=0.116.1",
    "httpx==0.25.2",
    "lxml>=6.0.0",
    "motor>=3.7.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pyjwt==2.8.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]

